# PythonApi/openapi.json
{"openapi": "3.1.0", "info": {"title": "Expense Tracker API", "description": "Smart Expense Tracker with automatic categorization", "version": "v1"}, "paths": {"/": {"get": {"summary": "Redirect To Swagger", "operationId": "redirect_to_swagger__get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/api/Expenses": {"get": {"tags": ["Expenses"], "summary": "Get Expenses", "operationId": "GetExpenses", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"items": {"$ref": "#/components/schemas/ExpenseItem"}, "type": "array", "title": "Response Getexpenses"}}}}}}, "post": {"tags": ["Expenses"], "summary": "Create Expense", "operationId": "CreateExpense", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/CreateExpenseCommand"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "integer", "title": "Response Createexpense"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/Expenses/categorize": {"post": {"tags": ["Expenses"], "summary": "Categorize Expense", "operationId": "CategorizeExpense", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/CategorizeExpenseRequest"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CategorizeExpenseResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/Expenses/{id}": {"delete": {"tags": ["Expenses"], "summary": "Delete Expense", "operationId": "DeleteExpense", "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}}, "components": {"schemas": {"CategorizeExpenseRequest": {"properties": {"description": {"type": "string", "title": "Description"}, "amount": {"anyOf": [{"type": "number"}, {"type": "string"}], "title": "Amount"}}, "type": "object", "required": ["description", "amount"], "title": "CategorizeExpenseRequest"}, "CategorizeExpenseResponse": {"properties": {"category": {"type": "string", "title": "Category"}, "confidence": {"type": "number", "title": "Confidence"}}, "type": "object", "required": ["category", "confidence"], "title": "CategorizeExpenseResponse"}, "CreateExpenseCommand": {"properties": {"raw_text": {"type": "string", "title": "Raw Text"}}, "type": "object", "required": ["raw_text"], "title": "CreateExpenseCommand"}, "ExpenseItem": {"properties": {"id": {"type": "integer", "title": "Id"}, "description": {"type": "string", "title": "Description"}, "amount": {"type": "string", "title": "Amount"}, "category": {"type": "string", "title": "Category"}, "date": {"type": "string", "format": "date-time", "title": "Date"}, "raw_text": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Raw Text"}}, "type": "object", "required": ["id", "description", "amount", "category", "date"], "title": "ExpenseItem"}, "HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}}}